chore: remove submitReview and related modules
- Delete submitReviewTool and publishReview modules
- Re-route summary coordination imports to publishSummaryOnly
- Retain publishReview harness helper inside test utilities
