+++ docs/development.md
@@ -8,7 +8,7 @@ Binding review rules live in [`.pr-agent/*.mdc`](../.pr-agent/) — this guide i
 
 | Area                      | Path                                           | Public entry                                                                                                                                                                                                                                 |
 | ------------------------- | ---------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| Review run + publish      | `src/review/`                                  | `orchestrator/orchestratorRun.ts`, `publish/publishReview.ts`, `ci/analyzeCi.ts` (LLM CI summary: `ci/authorCiSummary.ts`); metrics/footer helpers under `run/`                                                                              |
+| Review run + publish      | `src/review/`                                  | `orchestrator/orchestratorRun.ts`, `publish/publishSummaryOnly.ts`, `publish/publishFindingBatch.ts`, `ci/analyzeCi.ts` (LLM CI summary: `ci/authorCiSummary.ts`); metrics/footer helpers under `run/`                                       |
 | Local PR workspace        | `src/prWorkspace/`                             | `index.ts` (`withPrRepositoryView`); `workspaceResource.ts` owns temp-root allocation, ownership marker/heartbeat, credentials, and idempotent release                                                                                       |
 | Code index (optional FTS) | `src/codeIndex/`                               | `buildJob.ts`, `search.ts`, `repository.ts`                                                                                                                                                                                                  |
 | Agent work intake         | `src/agentWork/intake/`                        | `planner.ts` (pure), `applier.ts` (Postgres + pg-boss)                                                                                                                                                                                       |
