+++ test/helpers/reviewPublishTestHelpers.ts
@@ -1,18 +1,156 @@
-import { publishReview } from "../../src/review/publish/publishReview.js";
-import { prepareReviewPayloadForPublish } from "../../src/review/findings/findingPipeline.js";
-import type { EvidenceLedger } from "../../src/review/findings/evidenceLedger.js";
-import type { InlinePlacement } from "../../src/review/placement/reviewDiffPlacement.js";
-import type { ReviewFinding } from "../../src/review/reviewSchema.js";
+import type { Config } from "../../src/config.js";
+import { reviewCheckDetailsUrl } from "../../src/agentWork/reviewCheckRun.js";
 import type { AnyReviewLens } from "../../src/settings/legacyReviewLenses.js";
+import type { PrSurface } from "../../src/github/prSurface.js";
 import {
   createCachedPrDiffIndex,
   ingestListPullRequestFilesResult,
   type CachedPrDiffIndex,
 } from "../../src/review/placement/reviewDiffIndex.js";
-import { createSubmitReviewState } from "../../src/review/publish/submitReviewTool.js";
+import type { AcceptedPlacement } from "../../src/review/orchestrator/orchestratorTypes.js";
+import {
+  applyFindingLedgerDelta,
+  createFindingLedger,
+} from "../../src/review/orchestrator/orchestratorTypes.js";
+import type { CiSummaryAuthor } from "../../src/review/ci/authorCiSummary.js";
+import type { ReviewPayload, ReviewPublishContext } from "../../src/review/reviewSchema.js";
+import type { RepoPolicyResult } from "../../src/review/repoPolicy.js";
+import type { BoundPolicyJudge } from "../../src/review/publish/boundPolicyJudge.js";
+import { publishFindingBatch } from "../../src/review/publish/publishFindingBatch.js";
+import {
+  publishReviewSummaryOnly,
+  type RecordPublishStepWithCoordination,
+} from "../../src/review/publish/publishSummaryOnly.js";
+import { prepareReviewPayloadForPublish } from "../../src/review/findings/findingPipeline.js";
+import type { EvidenceLedger } from "../../src/review/findings/evidenceLedger.js";
+import type { InlinePlacement } from "../../src/review/placement/reviewDiffPlacement.js";
+import type { ReviewFinding } from "../../src/review/reviewSchema.js";
 import { createTestEvidenceLedger, seedEvidenceForFindings } from "./evidenceTestHelpers.js";
 
-/** Runs pre-publish pipeline then publishReview (matches submitReview path). */
+export type SubmitReviewState = {
+  published: boolean;
+  inlineReviewIds: number[];
+  threadCallCount: number;
+  lastValidationError: string | null;
+  publishCallCount: number;
+  publishCallsExhausted: boolean;
+  /** Set when publish is skipped (superseded, cancelled, or abort-check failed). */
+  publishSuperseded: boolean;
+};
+
+export function createSubmitReviewState(initial?: {
+  readonly published?: boolean;
+  readonly inlineReviewIds?: readonly number[];
+  readonly threadCallCount?: number;
+}): SubmitReviewState {
+  return {
+    published: initial?.published ?? false,
+    inlineReviewIds: [...(initial?.inlineReviewIds ?? [])],
+    threadCallCount: initial?.threadCallCount ?? 0,
+    lastValidationError: null,
+    publishCallCount: 0,
+    publishCallsExhausted: false,
+    publishSuperseded: false,
+  };
+}
+
+/** Legacy full publish path retained for publishReview.* unit tests. */
+export async function publishReview(
+  params: ReviewPublishContext & {
+    prSurface: PrSurface;
+    mode?: AnyReviewLens;
+    cfg: Pick<Config, "piModel" | "features">;
+    payload: ReviewPayload;
+    dedupedFindingCount?: number;
+    publishState: SubmitReviewState;
+    cachedDiffIndex?: CachedPrDiffIndex;
+    shouldLinkToSummary?: boolean;
+    progressCommentIdHint?: number | null;
+    staleReview?: boolean;
+    recordPublishStep?: RecordPublishStepWithCoordination;
+    storedInlineFingerprints?: readonly string[];
+    ciSummaryAuthor?: CiSummaryAuthor;
+    workItemId?: string;
+    resumedPlacements?: readonly AcceptedPlacement[];
+    shouldAbortPublish?: () => Promise<boolean>;
+    publishAbortState?: { readonly staleHead?: boolean };
+    repoPolicy?: RepoPolicyResult;
+    sameRepo?: boolean;
+    boundPolicyJudge?: BoundPolicyJudge;
+    readCheckoutFile?: (path: string) => Promise<string | undefined>;
+  },
+): Promise<void> {
+  const resumedPlacements = params.resumedPlacements ?? [];
+  let ledger = createFindingLedger({
+    accepted: resumedPlacements,
+    suppressionFingerprints: params.storedInlineFingerprints,
+    inlineReviewIds: params.publishState.inlineReviewIds,
+    postedInlineCount: resumedPlacements.length,
+    threadCallCount: params.publishState.threadCallCount,
+  });
+  const batchResult = await publishFindingBatch(params.payload.findings, {
+    ctx: params,
+    source: "review",
+    workItemId:
+      params.workItemId ?? params.recordPublishStep?.summaryCommentCoordination?.workItemId,
+    operationIntent: params.recordPublishStep?.summaryCommentCoordination
+      ? {
+          client: params.recordPublishStep.summaryCommentCoordination.pool,
+          workItemId: params.recordPublishStep.summaryCommentCoordination.workItemId,
+          resourceKey: params.recordPublishStep.summaryCommentCoordination.resourceKey,
+          leaseEpoch: params.recordPublishStep.summaryCommentCoordination.leaseEpoch,
+        }
+      : undefined,
+    resolveProgressCommentUrl: async () =>
+      reviewCheckDetailsUrl(
+        params.owner,
+        params.repo,
+        params.prNumber,
+        params.progressCommentIdHint,
+      ),
+    prSurface: params.prSurface,
+    cachedDiffIndex: params.cachedDiffIndex ?? createCachedPrDiffIndex(),
+    recordPublishStep: params.recordPublishStep,
+    shouldAbortPublish: params.shouldAbortPublish,
+    publishAbortState: params.publishAbortState,
+    repoPolicy: params.repoPolicy,
+    sameRepo: params.sameRepo,
+    boundPolicyJudge: params.boundPolicyJudge,
+    readCheckoutFile: params.readCheckoutFile,
+    ledger,
+  });
+  if (batchResult.kind === "stopped") {
+    params.publishState.publishSuperseded = true;
+    return;
+  }
+
+  ledger = applyFindingLedgerDelta(ledger, batchResult.delta);
+  params.publishState.inlineReviewIds = [...ledger.inlineReviewIds];
+  params.publishState.threadCallCount = ledger.threadCallCount;
+
+  const summaryResult = await publishReviewSummaryOnly({
+    cfg: params.cfg,
+    ctx: params,
+    prSurface: params.prSurface,
+    payload: params.payload,
+    ledger,
+    mode: params.mode,
+    cachedDiffIndex: params.cachedDiffIndex,
+    shouldLinkToSummary: params.shouldLinkToSummary,
+    progressCommentIdHint: params.progressCommentIdHint,
+    staleReview: params.staleReview,
+    recordPublishStep: params.recordPublishStep,
+    ciAuthor: params.ciSummaryAuthor,
+    shouldAbortPublish: params.shouldAbortPublish,
+    publishAbortState: params.publishAbortState,
+    dedupedFindingCount: params.dedupedFindingCount,
+  });
+  if (summaryResult.kind === "stopped") {
+    params.publishState.publishSuperseded = true;
+  }
+}
+
+/** Runs pre-publish pipeline then publishReview (legacy test harness). */
 export async function publishReviewForTest(
   params: Parameters<typeof publishReview>[0] & {
     mode?: AnyReviewLens;
